# optimize GRU by pre-calculating three linear operations