--- a/spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/LSTMPeephole.scala
+++ b/spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/LSTMPeephole.scala
@@ -67,30 +67,13 @@ class LSTMPeephole[T : ClassTag] (
   var cellLayer: Sequential[T] = _
   override var cell: AbstractModule[Activity, Activity, T] = buildLSTM()
 
-  override def preTopology: AbstractModule[Activity, Activity, T] =
-    Sequential()
-      .add(ConcatTable()
-        .add(Identity())
-        .add(Sequential()
-          .add(Dropout(p))
-          .add(TimeDistributed[T](Linear(inputSize, hiddenSize, wRegularizer = wRegularizer,
-            bRegularizer = bRegularizer)))))
-      .add(JoinTable(3, 1))
-
-  def buildGate(dimension: Int, offset: Int, length: Int, preTopo: Boolean = true):
-  Sequential[T] = {
+  def buildGate(): Sequential[T] = {
     val gate = Sequential()
 
     val i2g = Sequential()
-    if (preTopo) {
-      i2g.add(Narrow(dimension, offset, length))
-    } else {
-      i2g
-        .add(Narrow(dimension, offset, length))
-        .add(Dropout(p))
-        .add(Linear(inputSize, hiddenSize, wRegularizer = wRegularizer,
-          bRegularizer = bRegularizer))
-    }
+      .add(Dropout(p))
+      .add(Linear(inputSize, hiddenSize, wRegularizer = wRegularizer,
+        bRegularizer = bRegularizer))
     val h2g = Sequential()
       .add(Dropout(p))
       .add(Linear(hiddenSize, hiddenSize,
@@ -106,17 +89,17 @@ class LSTMPeephole[T : ClassTag] (
   }
 
   def buildInputGate(): Sequential[T] = {
-    inputGate = buildGate(2, 1, inputSize, false)
+    inputGate = buildGate()
     inputGate
   }
 
   def buildForgetGate(): Sequential[T] = {
-    forgetGate = buildGate(2, 1 + inputSize, hiddenSize, true)
+    forgetGate = buildGate()
     forgetGate
   }
 
   def buildOutputGate(): Sequential[T] = {
-    outputGate = buildGate(2, 1, inputSize, false)
+    outputGate = buildGate()
     outputGate
   }
 
@@ -125,7 +108,6 @@ class LSTMPeephole[T : ClassTag] (
       .add(NarrowTable(1, 2))
 
     val i2h = Sequential()
-      .add(Narrow(2, 1, inputSize))
       .add(Dropout(p))
       .add(Linear(inputSize, hiddenSize, wRegularizer = wRegularizer,
         bRegularizer = bRegularizer))