## What changes were proposed in this pull request?
define preTopology in GRU

## How was this patch tested?
Unit Test.

Since the structure of GRU in BigDL and Torch are different, I revised the Unit Test by rearrange the parameters in copy operation.




The speed:

I set batchSize = 12, inputSize = 1000, hiddenSize = 1000, and time = 200.
Running on my Desktop.
```
elapsed time | before    | after
------------------------------------------
1                | 3.22405 | 2.91554
2                | 3.26716 | 2.94787
3                | 3.17547 | 2.96847
4                | 3.15196 | 2.99957
5                | 3.16941 | 2.93977
------------------------------------------
average        | 3.19761 | 2.95424
------------------------------------------
throughput    | 3.7528  | 4.06195 
```
Thus, the revised version will give a ~8.24% throughput improvement.
With larger batchSize, inputSize and hiddenSize, the improvement will be more salient.