@jason-dai 2017-07-05
Shall we do something similar for LSTM?