@JerryYanWan 2017-07-05
The LSTM has already been optimized for input2hidden calculation.
I'm working on LSTMPeephole with similar strategy. Will propose PR soon.